Move AffineStar into its own module.
